feat: update README.md and change context of EE experimentation
Due to increasing difficulties with OSM and OSMnx, make the decision to
change the context of the EE experimental work to that of the tram
network in Vienna. The aims of the EE, however, still remain the same.
As a result, construct new graph for tram network. Implement
functionality to vary input size of graph, run Dijkstra's
algorithm on graph, and plot graph visually.