feat: change scenario to that of the 15-puzzle game, and change RQ
Change scenario from Vienna tram graph to that of London buses. Then
change scenario to that of the 15-puzzle game (now graph represents game
states rather than geographical connections). RQ has also changed to "To
what extent does the search depth impact the quality of different
heuristics of the A* search algorithm in solving the single-pair
shortest path problem in an unweighted graph?"